PENGSOL-483: Update the ReadMe file. Cleanup. Split one file into a proper file structure. Add common mistakes handling.
